Removed numeric limits, fixed a vector out of range bug
Numeric limits for unsigned int didn't seem to work as intended, causing
the algorithm to not yield any successfull results. Infinity of unsigned
int returns 0 as it can not have infinity. Reverting to UINT32_MAX.

Reverted a bug where the path vector was no longer iterated over
correctly resulting in an out of bounds exception.